* commit 'c49fc1c773033a24a6dcc2efd8b5b8bcccd0703c':
  changes in adding rtr_status and rtr_amount
  changes in adding rtr_status and rtr_amount
  changes in adding rtr_amount and rtr_status in rate details table
  Changes in adding rtr_amount and rtr_status in audit rate table.
  changes in capturing rtr_amount and rtr_status in rate details table.
  changes in proc to sql query in DAO

+++ src/main/java/com/envista/msi/api/web/rest/util/audit/parcel/ParcelRateRequestBuilder.java
@@ -149,7 +149,7 @@ public class ParcelRateRequestBuilder {
                     dimensionsObj.setUnits(dimUnit);
 
                     ParcelRateRequest.Item item = new ParcelRateRequest.Item();
-                    item.setSequence(latestFreightCharge.getParentId().intValue());
+                    item.setSequence(latestFreightCharge.getParentId());
                     item.setWeight(weightObj);
                     item.setActualWeight(actualWeightElement);
                     item.setQuantity(quantityObj);
@@ -341,7 +341,7 @@ public class ParcelRateRequestBuilder {
                         dimensionsObj.setUnits(dimUnit);
 
                         ParcelRateRequest.Item item = new ParcelRateRequest.Item();
-                        item.setSequence(firstBaseCharge.getParentId().intValue());
+                        item.setSequence(firstBaseCharge.getParentId());
                         item.setWeight(weightObj);
                         item.setActualWeight(actualWeightElement);
                         item.setQuantity(quantityObj);
